format exception-checker-spring-scenario and exception-checker-tomcat-scenario
