+++ test/plugin/scenarios/exception-checker-spring-scenario/src/main/java/test/apache/skywalking/apm/testcase/exceptionchecker/exception/TestAnnotatedException.java
@@ -16,29 +16,32 @@
  *
  */
 
-package test.org.apache.skywalking.apm.testcase.exceptionchecker.exception;
+package test.apache.skywalking.apm.testcase.exceptionchecker.exception;
 
-public class TestHierarchyListedException extends TestListedException {
+import org.apache.skywalking.apm.toolkit.trace.IgnoredException;
 
-    public TestHierarchyListedException() {
+@IgnoredException
+public class TestAnnotatedException extends RuntimeException {
+
+    public TestAnnotatedException() {
     }
 
-    public TestHierarchyListedException(final String message) {
+    public TestAnnotatedException(final String message) {
         super(message);
     }
 
-    public TestHierarchyListedException(final String message, final Throwable cause) {
+    public TestAnnotatedException(final String message, final Throwable cause) {
         super(message, cause);
     }
 
-    public TestHierarchyListedException(final Throwable cause) {
+    public TestAnnotatedException(final Throwable cause) {
         super(cause);
     }
 
-    public TestHierarchyListedException(final String message,
-                                        final Throwable cause,
-                                        final boolean enableSuppression,
-                                        final boolean writableStackTrace) {
+    public TestAnnotatedException(final String message,
+                                  final Throwable cause,
+                                  final boolean enableSuppression,
+                                  final boolean writableStackTrace) {
         super(message, cause, enableSuppression, writableStackTrace);
     }
 }
